final_project/final_project/demo segmentation/requirements.txt
albumentations==1.3.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.7.0
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
huggingface-hub==0.20.1
idna==3.6
imageio==2.33.1
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.6.2
natsort==8.4.0
networkx==3.1
numpy==1.23.5
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pathtools==0.1.2
Pillow==9.3.0
pip==23.3.2
promise==2.3
protobuf==4.25.1
psutil==5.9.7
ptflops==0.7.2.1
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==1.39.1
setproctitle==1.3.3
setuptools==68.2.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
torchinfo==1.8.0
tqdm==4.64.1
typing_extensions==4.9.0
urllib3==2.1.0
wandb==0.13.5
wheel==0.42.0
